refactor(filter): Clean up op mod
